requirements.txt
coverage==7.13.5
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.1.0
python-dotenv==1.0.0
